Muevo funcionalidad de testeo de envío de mails a la configuración del admin
