+++ ckanext/gobar_theme/js/commons.js
