+++ ckanext/gobar_theme/templates/user/login.html
@@ -43,7 +43,7 @@
     </div>
 </div>
 
-{% resource 'gobar_js/user_config/user_commons.js' %}
+{% resource 'gobar_js/commons.js' %}
 {% resource 'gobar_js/login.js' %}
 
 {% endblock %}
