+++ ckanext/gobar_theme/templates/user/perform_reset.html
@@ -47,7 +47,7 @@
     </div>
 
 
-    {% resource 'gobar_js/user_config/user_commons.js' %}
+    {% resource 'gobar_js/commons.js' %}
     {% resource 'gobar_js/user/secure_password.js' %}
     {% resource 'gobar_js/user/password_reset.js' %}
 {% endblock %}
