+++ ckanext/gobar_theme/templates/user/user_config_create_users.html
@@ -104,8 +104,8 @@
 
     </div>
     {% snippet 'snippets/csrf.html' %}
+    {% resource 'gobar_js/commons.js' %}
     {% resource 'gobar_js/user/secure_password.js' %}
-    {% resource 'gobar_js/user_config/user_commons.js' %}
     {% resource 'gobar_js/user_config/create_users.js' %}
     {% resource 'gobar_js/user_config/multiple-select.js' %}
 {% endblock %}
